Checkpoints: move utils back to app
